Same functionality not a api, Kafka communication w/ ExpMangFunc

+++ kubeController/kubeController.py
@@ -1,10 +1,21 @@
+'''
+	Rodrigo Santos, n mec 89180
+	
+	- Dissertation Project
+	Computation Controller - Kubernetes Controller
+'''
+
 from kubernetes import config, dynamic
 from kubernetes.client import api_client
-from fastapi import FastAPI
+from kafka import KafkaConsumer
+from kafka import KafkaProducer
+
 import kubernetes.client
 import datetime
 import pytz
-import p4runtime
+import subprocess
+import requests
+import sys
 
 # namespace test: "default"
 # name test: "nginx-deployment"
@@ -15,107 +26,218 @@ import p4runtime
 # name service test: "frontend-service"
 # protocol test: "TCP"
 
+#app = FastAPI()
+
 global client
+global networking_v1_api
+global consumer
+global producer
+global expMngFncAPI
 
+'''configuration = kubernetes.client.Configuration()
+configuration.api_key['authorization'] = '754quh.kakrr1jsc4isb4oz'
+configuration.api_key_prefix['authorization'] = 'Bearer'
+configuration.host = "http://10.0.2.15"
+configuration.verify_ssl = False
+
+client = dynamic.DynamicClient(
+        api_client.ApiClient(configuration)
+)'''
+
+consumer = KafkaConsumer("ComputationManagment", bootstrap_servers='10.0.2.15:9092')
+
+producer = KafkaProducer(bootstrap_servers='10.0.2.15:9092')
+
+client = dynamic.DynamicClient(
+        api_client.ApiClient(configuration=config.load_kube_config(config_file='config/config'))
+)
+
+networking_v1_api = kubernetes.client.NetworkingV1Api()
+
+
+###############################################
+#		Main Cycle			#
+###############################################
 def main():
-    # Creating a dynamic client
-    global client
-    
-    client = dynamic.DynamicClient(
-        api_client.ApiClient(configuration=config.load_kube_config())
-    )
-    networking_v1_api = kubernetes.client.NetworkingV1Api()
-    
-    op = '99'
-    while op != '0':
-    	print("\n -> KUBE CONTROLLER:\n" +
-    		"\t1 - Create Deployment\n" +
-    		"\t2 - Delete Deployment\n" +
-    		"\t3 - Create Service\n" +
-    		"\t4 - Delete Service\n" +
-    		"\t5 - Create Ingress\n" +
-    		"\t6 - Delete Ingress\n" +
-    		"\t7 - List Pods\n" +
-    		"\t8 - List Services\n" +
-    		"\t9 - List Ingresses\n" +
-    		"\t0 - Exit\n")
-    	op = input("[OPTION] -> ")
 
-    	if op == '1':
-    	    create_deployment()
-    	if op == '2':
-    	    delete_deployment()
-    	if op == '3':
-    	    create_service()
-    	if op == '4':
-    	    delete_service()
-    	if op == '5':
-    	    create_ingress(networking_v1_api)
-    	if op == '6':
-    	    delete_ingress(networking_v1_api)
-    	if op == '7':
-    	    list_pods()
-    	if op == '8':
-    	    list_services()
-    	if op == '9':
-    	    list_ingresses(networking_v1_api)
+    global consumer
+    
+    while(True):
+        for msg in consumer:
+            processed_msg = msg.value.decode()
+            log(processed_msg)
+            checkAction(processed_msg)
+            processed_msg = ''
 
-    print('\n[EXITING]\n')
+
+###############################################
+#		Check Action			#
+###############################################
+def checkAction(msg):
+
+    global producer
+    global expMngFncAPI
+    
+    processor, action, mode = msg.split(' ')[0:3]
+    
+    if processor != '[COMPUTATIONCONTROLLER]':
+        return None
+
+    ## INSTANTIATIONS ------
+    if action == '[INSTANTIATE]':
+        
+        ## RIGHTAWAY MODE ------
+        if mode == '[RIGHTAWAY]':
+            log(" - creating deployment -")
+            namespace, name, app, container_name, image = msg.split(' ')[3:]
+            res_dep = create_deployment(namespace, name, app, container_name, image)
+        
+            # Create service
+            log(" - creating service -")
+            res_ser = create_service(name= name+"-service", selector=app, protocol="TCP", port = 5000, targetPort = 50000)
+        
+            if res_dep == {"SUCCESS"} and res_ser == {"SUCCESS"}:
+                msg_out = '[NETCONTROLLER] [INSERT] IPV4SubEntry 10.30.0.30 3 10.0.2.15 08:00:27:93:75:80'
+                producer.send('NetManagment', msg_out.encode())
+            else:
+                msg_out = '[MANAGMENT] [ERROR] [DELETE]'
+                producer.send('ComputationManagment', msg_out.encode())
+                
+                
+        ## TRIGGERBASED MODE ------
+        if mode == '[TRIGGERED]':
+            log(" - fetching deployment information -")
+            
+            port = msg.split(' ')[3:]
+            info = requests.get("http://"+expMngFncAPI+":8000/fetchInfo?port=1").content.decode()
+            
+            namespace, name, app, container_name, image = info.replace('"', '').replace('[', '').replace(']','').split(",")
+
+            log(" - creating deployment -")
+            res_dep = create_deployment(namespace, name, app, container_name, image)
+        
+            # Create service
+            log(" - creating service -")
+            res_ser = create_service(name= name+"-service", selector=app, protocol="TCP", port = 5000, targetPort = 50000)
+        
+            if res_dep == {"SUCCESS"} and res_ser == {"SUCCESS"}:
+                msg_out = '[NETCONTROLLER] [INSERT] IPV4SubEntry 10.30.0.30 3 10.0.2.15 08:00:27:93:75:80'
+                producer.send('NetManagment', msg_out.encode())
+            else:
+                msg_out = '[MANAGMENT] [ERROR] [DELETE]'
+                producer.send('ComputationManagment', msg_out.encode())
+                
+
+    ## DELETIONS ------
+    if action == '[DELETE]':
+    
+        ## RIGHTAWAY MODE ------
+        if mode == '[RIGHTAWAY]':
+            log(" - deleting deployment -")
+            namespace, name = msg.split(' ')[3:]
+            res_dep = delete_deployment(namespace, name)
+        
+            # Delete service
+            log(" - deleting service -")
+            res_ser = delete_service(name="server-udp-service")
+        
+            if res_dep == {"SUCCESS"} and res_ser == {"SUCCESS"}:
+                msg_out = '[NETCONTROLLER] [DELETE] IPV4SubEntry 10.30.0.30 3 10.0.2.15 08:00:27:93:75:80'
+                producer.send('NetManagment', msg_out.encode())
+            else:
+                msg_out = '[MANAGMENT] [ERROR] [DELETE]'
+                producer.send('ComputationManagment', msg_out.encode())
+                
+        ## TRIGGERBASED MODE ------
+        if mode == '[TRIGGERED]':
+            log(" - fetching deployment information -")
+            
+            port = msg.split(' ')[3:]
+            info = requests.get("http://"+expMngFncAPI+":8000/fetchInfo?port=1").content.decode()
+            
+            if info != None and info != "null":
+                namespace, name, app, container_name, image = info.replace('"', '').replace('[', '').replace(']','').split(",")
+            
+                log(" - deleting deployment -")
+                res_dep = delete_deployment(namespace, name)
+            
+                # Delete service
+                log(" - deleting service -")
+                res_ser = delete_service(name="server-udp-service")
+            
+                if res_dep == {"SUCCESS"} and res_ser == {"SUCCESS"}:
+                    msg_out = '[NETCONTROLLER] [DELETE] IPV4SubEntry 10.30.0.30 3 10.0.2.15 08:00:27:93:75:80'
+                    producer.send('NetManagment', msg_out.encode())
+                else:
+                    msg_out = '[MANAGMENT] [ERROR] [DELETE]'
+                    producer.send('ComputationManagment', msg_out.encode())
+            else:
+                msg_out = '[MANAGMENT] [ERROR] [DELETE] Nothing to delete!'
+                producer.send('ComputationManagment', msg_out.encode())
+
+
+###############################################
+#			LOG			#
+###############################################
+def log(msg):
+
+	# Building the command.
+	command = ['echo', msg]
+
+	subprocess.call(command)
 
 
 ###############################################
 #		Create Deployemnt		#
 ###############################################
-def create_deployment():
+def create_deployment(namespace="default", name="None", app="None", container_name="None", image="None"):
 
     global client
-    
+
+    if name=="None" or app=="None" or container_name=="None" or image=="None":
+        return {"ERROR"}
+
     # fetching the deployment api
     api = client.resources.get(api_version="apps/v1", kind="Deployment")
 
     deployment = ''
 
-    while deployment == '':
-        try:
-            namespace = input("\n\t -> Namespace: ")
-            name = input("\t -> Name: ")
-            app = input ("\t -> App: ")
-            container_name = input("\t -> Container name: ")
-            image = input("\t -> Image: ")
+    try:
             
-            
-            deployment_manifest = {
-                "apiVersion": "apps/v1",
-                "kind": "Deployment",
-                "metadata": {"labels": {"app": app}, "name": name},
-                "spec": {
-                    "replicas": 3,
-                    "selector": {"matchLabels": {"app": app}},
-                    "template": {
-                        "metadata": {"labels": {"app": app}},
-                        "spec": {
-                            "containers": [
-                                {
-                                    "name": container_name,
-                                    "image": image,
-                                    "ports": [{"containerPort": 80}],
-                                }
-                            ]
-                        },
+        deployment_manifest = {
+            "apiVersion": "apps/v1",
+            "kind": "Deployment",
+            "metadata": {"labels": {"app": app}, "name": name},
+            "spec": {
+                "replicas": 1,
+                "selector": {"matchLabels": {"app": app}},
+                "template": {
+                    "metadata": {"labels": {"app": app}},
+                    "spec": {
+                        "containers": [
+                            {
+                                "name": container_name,
+                                "image": image,
+                                "ports": [{"containerPort": 80}],
+                                "imagePullPolicy": "Never",
+                            }
+                        ],
+                        "restartPolicy": "Always",
                     },
                 },
-            }
+            },
+        }
 
-            # Creating deployment `nginx-deployment` in the `default` namespace
-            deployment = api.create(body=deployment_manifest, namespace=namespace)
-        except:
-            deployment = ''
-            print('\n[ERROR] Something went wrong!\n')
-        finally:    
-            print("\n[INFO] deployment `"+ name +"` created\n")
+        # Creating deployment `nginx-deployment` in the `default` namespace
+        deployment = api.create(body=deployment_manifest, namespace=namespace)
+    except:
+        deployment = ''
+        print('\n[ERROR] Something went wrong!\n')
+        return {"ERROR"}    
+        
 
     # Listing deployment `nginx-deployment` in the `default` namespace
-    deployment_created = api.get(name=name, namespace=namespace)
+    '''deployment_created = api.get(name=name, namespace=namespace)
 
     print("%s\t%s\t\t\t%s\t%s" % ("NAMESPACE", "NAME", "REVISION", "RESTARTED-AT"))
     print(
@@ -126,74 +248,51 @@ def create_deployment():
             deployment_created.metadata.annotations,
             deployment_created.spec.template.metadata.annotations,
         )
-    )
+    )'''
 
-    # Patching the `spec.template.metadata` section to add `kubectl.kubernetes.io/restartedAt` annotation
-    # In order to perform a rolling restart on the deployment `nginx-deployment`
-
-    deployment_manifest["spec"]["template"]["metadata"] = {
-        "annotations": {
-            "kubectl.kubernetes.io/restartedAt": datetime.datetime.utcnow()
-            .replace(tzinfo=pytz.UTC)
-            .isoformat()
-        }
-    }
-
-    deployment_patched = api.patch(
-        body=deployment_manifest, name=name, namespace=namespace
-    )
-
-    print("\n[INFO] deployment `"+ name +"` restarted\n")
-    print(
-        "%s\t%s\t\t\t%s\t\t\t\t\t\t%s"
-        % ("NAMESPACE", "NAME", "REVISION", "RESTARTED-AT")
-    )
-    print(
-        "%s\t\t%s\t%s\t\t%s\n"
-        % (
-            deployment_patched.metadata.namespace,
-            deployment_patched.metadata.name,
-            deployment_patched.metadata.annotations,
-            deployment_patched.spec.template.metadata.annotations,
-        )
-    )
+    print("\n[INFO] deployment `"+ name +"` created\n")
+    return {"SUCCESS"}
 
 
 ###############################################
 #		Delete Deployemnt		#
 ###############################################
-def delete_deployment():
+def delete_deployment(namespace="default", name="None"):
     
     global client
     
+    if name == "None":
+       return {"ERROR"} 
     # fetching the deployment api
     api = client.resources.get(api_version="apps/v1", kind="Deployment")
     
     # Deleting deployment `nginx-deployment` from the `default` namespace
     try:
-       namespace = input("\n\t -> Namespace: ")
-       name = input("\t -> Name: ")
        deployment_deleted = api.delete(name=name, body={}, namespace=namespace)
+       print("\n[INFO] deployment `"+ name +"` deleted\n")
+       return {"SUCCESS"}
     except:
        print("\n[ERROR] Something went wrong!")
-    finally:
-       print("\n[INFO] deployment `"+ name +"` deleted")
+       return {"ERROR"}
 
 
 ###############################################
 #		Create Service 		#
 ###############################################
-def create_service():
+def create_service(namespace="default", name="None", selector="None", protocol="None", port: int = 0, targetPort: int = 0):
 
     global client
+
+    if name=="None" or protocol=="None" or port==0 or targetPort==0:
+        {"ERROR"}
 
     # fetching the service api
     api = client.resources.get(api_version="v1", kind="Service")
 
     try:
-        namespace = input("\n\t -> Namespace: ")
-        name = input("\t -> Name: ")
-        protocol = input("\t -> Protocol: ")
+        #namespace = input("\n\t -> Namespace: ")
+        #name = input("\t -> Name: ")
+        #protocol = input("\t -> Protocol: ")
 
         service_manifest = {
             "apiVersion": "v1",
@@ -201,9 +300,9 @@ def create_service():
             "metadata": {"labels": {"name": name}, "name": name, "resourceversion": "v1"},
             "spec": {
                 "ports": [
-                    {"name": "port", "port": 80, "protocol": protocol, "targetPort": 80}
+                    {"name": "port", "port": port, "protocol": protocol, "targetPort": targetPort}
                 ],
-                "selector": {"name": name},
+                "selector": {"name": selector},
             },
         }
 
@@ -224,47 +323,44 @@ def create_service():
 
         # Patching the `spec` section of the `frontend-service`
 
-        service_manifest["spec"]["ports"] = [
-            {"name": "new", "port": 8080, "protocol": protocol, "targetPort": 8080}
-        ]
+        #service_manifest["spec"]["ports"] = [
+        #    {"name": "new", "port": 8080, "protocol": protocol, "targetPort": 8080}
+        #]
 
-        service_patched = api.patch(body=service_manifest, name=name, namespace=namespace)
+        #service_patched = api.patch(body=service_manifest, name=name, namespace=namespace)
 
     except:
         print("[ERROR] something went wrong!")
-    finally:
-        print("\n[INFO] service `"+ name + "` patched\n")
-        print("%s\t%s\t\t\t%s" % ("NAMESPACE", "NAME", "PORTS"))
-        print(
-            "%s\t\t%s\t%s\n"
-            % (
-                service_patched.metadata.namespace,
-                service_patched.metadata.name,
-                service_patched.spec.ports,
-            )
-        )
+        return {"ERROR"}
+    
+    return {"SUCCESS"}
 
 
 ###############################################
 #		Delete Service 		#
 ###############################################
-def delete_service():
+def delete_service(namespace="default", name="None"):
 
     global client
+
+    if name == "None":
+       return {"ERROR"}
 
     # fetching the service api
     api = client.resources.get(api_version="v1", kind="Service")
 
     try:
-       namespace = input("\n\t -> Namespace: ")
-       name = input("\t -> Name: ")
+       #namespace = input("\n\t -> Namespace: ")
+       #name = input("\t -> Name: ")
 
        # Deleting service `frontend-service` from the `default` namespace
        service_deleted = api.delete(name=name, body={}, namespace=namespace)
+       print("\n[INFO] service `" + name + "` deleted\n")
     except:
        print("\n[ERROR] something went wrong!")
-    finally:
-       print("\n[INFO] service `" + name + "` deleted\n")
+       return {"ERROR"}
+       
+    return {"SUCCESS"}
 
 
 ###############################################
@@ -364,5 +460,9 @@ def list_ingresses(networking_v1_api):
         print("\n[INFO] no ingresses to show!")
 
 
+################################################
 if __name__ == "__main__":
+    for i in range(1, len(sys.argv),2):
+        if (sys.argv[i] == "--expMngFncAPI" or sys.argv[i] == "-api") and i != len(sys.argv) - 1:
+            expMngFncAPI = sys.argv[i + 1]
     main()
